requirements.txt
whoosh
pytesseract
pdfminer.six
python-docx
pillow
tk
